% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ipwsvm.R
\name{fit.unadj.svm}
\alias{fit.unadj.svm}
\title{Fitting stuff}
\usage{
fit.unadj.svm(tr.x, tr.y, ts.x, ts.y, grid.C)
}
\arguments{
\item{tr.x}{Matrix or data.frame of features}

\item{tr.y}{Vector of group labels}

\item{ts.x}{Matrix or data.frame of features for testing}

\item{ts.y}{Vector of group labels in test set}

\item{grid.C}{Vector of C values for tuning the SVM}
}
\description{
Fit unweighted SVM
}
\examples{
if (require(MASS)){
##################################################
# Generate data
##################################################
set.seed(5000)
n = 500
p = 2
sigma = matrix(.25, p, p)
diag(sigma) = 1
eps1 = mvrnorm(n, rep(0, p), sigma)
eps2 = mvrnorm(n, rep(0, p), sigma)
d = rbinom(n, 1, .5)
x1 = 5 - .25*d + .2*d*eps1[,1] + .1*eps2[,1]
x2 = 5 - d + .3*eps2[,1] + .5*(eps1[,2] + eps2[,2])
par(mfrow=c(1,1))
plot(x1, x2, col=d+2)
X = cbind(x1, x2)
ts.eps1 = mvrnorm(n, rep(0, p), sigma)
ts.eps2 = mvrnorm(n, rep(0, p), sigma)
ts.d = rbinom(n, 1, .5)
ts.x1 = 5 - .25*ts.d + .2*ts.d*ts.eps1[,1] + .1*ts.eps2[,1]
ts.x2 = 5 - ts.d + .3*ts.eps2[,1] + .5*(ts.eps1[,2] + ts.eps2[,2])
par(mfrow=c(1,1))
plot(ts.x1, ts.x2, col=ts.d+2)
ts.X = cbind(ts.x1, ts.x2)
tr.x = X
tr.y = d
ts.x = ts.X
ts.y = ts.d
grid.C = 10^c(-3:3)
out = fit.unadj.svm(tr.x = tr.x, tr.y = tr.y, ts.x = ts.x, ts.y = ts.y, grid.C = grid.C)
print(out)
}
}
\author{
Kristin Linn
}
\seealso{
\code{data.frame}, \code{\link{data.frame}}
}

